split into multiple files and move into solver package
