+++ src/fenics_constitutive/solver/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from ._solver import IncrSmallStrainProblem
+
+__all__ = ["IncrSmallStrainProblem"]
